Import `sslkeylogfile` configuration
This adds the `sslkeylog` configuration option.  If enabled during
build and the SSLKEYLOGFILE env var is set, all secrets are recorded
to that file. *Do not enable or use this in production.*
Ported from OpenSSL PR#25297

Also, some minor doc fixes, mainly to remove references to
commands and variables that were missed during PR#101 and PR#135.

Co-authored-by: Theo Buehler <botovq@users.noreply.github.com>