Add -fno-sanitize=function
This does not report indirect calls of a function through a function
pointer of the wrong type. This UB pattern is present at least in the
stack and lhash code and should be fixed. BoringSSL has fixes, but it
isn't pretty. Until a solution is chosen, we want CI to work.